Add files via upload
The application captures an image from the user's webcam and uses a pre-trained DeepFace model to predict the emotion expressed in the captured image. The supported emotions are angry, disgust, fear, happy, sad, surprise, and neutral.The user interface is created using Streamlit, making it easy for users to interact with the application. It displays a "Capture Emotion" button that initiates the webcam capture process.