Add files via upload
This Python script is a Streamlit web application that allows users to perform real-time emotion detection using a webcam, upload an image from their system for emotion detection, or capture an image from the webcam and analyze the emotion detected in the image. Here's a description of the main features and functionality:

Emotion Detection Options:

The application provides three options in the sidebar for users to choose from:
"Real-time Emotion Detection": Analyzes emotions in real-time using the webcam.
"Capture Image from System": Allows users to upload an image from their system for emotion analysis.
"Click Image": Captures an image using the webcam and analyzes the detected emotion.
Real-time Emotion Detection:

When the "Real-time Emotion Detection" option is selected, the webcam is activated, and the application continuously captures video frames.
It processes each frame to detect the emotion of the face in the frame using a pre-trained deep learning model (DEEPFACE).
The detected emotion is displayed in real-time along with a rectangle drawn around the face and a caption.
Capture Image from System:

Users can upload an image (in formats such as JPG, JPEG, or PNG) from their local system using the file uploader.
The uploaded image is processed, and the detected emotion is displayed, along with the uploaded image.
Based on the detected emotion, a corresponding response or message is displayed.
Click Image:

By clicking the "Click Image" button, users can capture an image using their webcam.
Similar to the "Capture Image from System" option, the detected emotion is displayed along with the captured image.
A response or message based on the detected emotion is also provided.
Emotion and Response:

The application recognizes emotions such as "happy," "sad," "angry," "neutral," "surprise," "disgust," and "fear."
For each detected emotion, a predefined response or message is displayed to provide a user-friendly interaction.
Text-to-Speech:

The application utilizes the pyttsx3 library to convert the response or message into speech for a more interactive experience.
User Interaction:

Users can choose their preferred option from the sidebar, upload images, and interact with the application through button clicks.
Graceful Exit:

For the real-time emotion detection option, users can press 'q' to exit the webcam video feed and close the application.
Overall, this Streamlit application offers an engaging and interactive way for users to explore real-time emotion detection and image analysis using pre-trained deep learning models while providing responses and messages based on the detected emotions.